[FIX] stock_picking_invoice_link: don't store auxiliar fields
In c332869b09d99231f3c5f957089fddd257ee23ba a new computed field was
introduced to show the count of a move's related pickings. But it was
made stored, which is needless for this kind of field and adds a huge
overhead when migrating from big DBs from previous versions.

With this commit:

- We drop the storing of this field which is meant to play a UI role.
- We rename the field to picking count, as a more precise nomenclature.

TT46020